+++ lib/src/connection_pool.dart
@@ -49,6 +49,15 @@ class SqliteConnectionPool with SqliteQueries implements SqliteConnection {
       : _writeConnection = writeConnection,
         _upstreamPort = upstreamPort;
 
+  /// Returns true if the _write_ connection is currently in autocommit mode.
+  @override
+  Future<bool> getAutoCommit() async {
+    if (_writeConnection == null) {
+      throw AssertionError('Closed');
+    }
+    return await _writeConnection!.getAutoCommit();
+  }
+
   @override
   Future<T> readLock<T>(Future<T> Function(SqliteReadContext tx) callback,
       {Duration? lockTimeout, String? debugContext}) async {