+++ lib/src/sqlite_database.dart
@@ -109,6 +109,13 @@ class SqliteDatabase with SqliteQueries implements SqliteConnection {
     return _pool.closed;
   }
 
+  /// Returns true if the _write_ connection is in auto-commit mode
+  /// (no active transaction).
+  @override
+  Future<bool> getAutoCommit() {
+    return _pool.getAutoCommit();
+  }
+
   void _listenForEvents() {
     UpdateNotification? updates;